convert

+++ part5/lesson12/src/test/java/ru/amizichenko/tracker/services/IteratorEvenNumbersTest.java
@@ -14,7 +14,7 @@ public class IteratorEvenNumbersTest {
     public void whenGetNextCallShouldReturnEvenNumberAndPointerForward() {
         IteratorEvenNumbers it = new IteratorEvenNumbers(new int[] {1,1,1,2,1,1,4});
         it.next();
-        int resoult = (Integer) it.next();
+        int resoult = it.next();
 
         assertThat(resoult, is(4));
     }
